+++ app/Http/Controllers/Classes/StudioController.php
@@ -2,11 +2,13 @@
 
 namespace App\Http\Controllers\Classes;
 
-use App\Http\Controllers\Controller;
-use Illuminate\Http\Request;
-use App\Models\Studio;
-use Yajra\DataTables\Facades\DataTables;
 use Carbon\Carbon;
+use App\Models\Studio;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Log;
+use App\Http\Controllers\Controller;
+use Yajra\DataTables\Facades\DataTables;
 
 class StudioController extends Controller
 {
@@ -51,13 +53,35 @@ class StudioController extends Controller
     /**
      * Remove the specified resource from storage.
      */
+    
     public function destroy(string $id)
     {
-        $studio = Studio::find($id);
-        $studio->delete();
-        return response()->json([
-            'status' => true,
-            'message' => 'Đã xóa thành công',
-        ], 200);
+        try {
+            DB::beginTransaction();
+        
+            $studio = Studio::find($id);
+        
+            if ($studio) {
+                $studio->delete();
+        
+                DB::table('movie_studio')->where('studio_id', $id)->delete();
+        
+                DB::commit();
+                return response()->json([
+                    'status' => true,
+                    'message' => 'Đã xóa thành công',
+                ], 200);
+            } else {
+                throw new \Exception("Studio không tồn tại.");
+            }
+        
+        } catch (\Exception $e) {
+            DB::rollBack();
+            Log::emergency("File:" . $e->getFile() . " Line:" . $e->getLine() . " Message:" . $e->getMessage());
+            return response()->json([
+                'status' => false,
+                'message' => $e->getMessage(),
+            ], 201);
+        }
     }
 }
